Fix UPF nodeID parsing issue when using FQDN
When an FQDN is used as UPF node_id in SMF configuration, IP parsing fails 
and PFCP session is not established between SMF and UPF. To resolve this 
issue, and since there is no native way to check whether node_is is IPv4, IPv6 
or an FQDN, we need to try a name resolution before parsing the IP.

This issue was experienced in the docker version of free5gc: https://github.com/free5gc/free5gc-compose/issues/9
Thanks to @hhoai we now know the root cause of the problem and
the solution. The verification step will make sure that the nodeID is
updated only when the DNS resolution succeeds.